# api/openapi-combined.yaml
openapi: 3.0.3
info:
  title: Git Simulation Platform API
  description: |
    Полное API платформы для моделирования Git с Code Review.
    
    ## Архитектура
    
    Платформа состоит из трёх микросервисов:
    
    ### 1. User Gateway Service (порт 8082)
    Фасад-сервис для конечных пользователей. Оркестрирует остальные сервисы,
    валидирует права доступа. **Рекомендуется использовать этот сервис.**
    
    Все запросы используют **имена** (team_name, username, repo_name, commit_name, pr_name).
    UUID/ID возвращаются только в ответах для справки.
    
    ### 2. PR Allocation Service (порт 8080)  
    Управление командами, пользователями и Pull Requests.
    Автоматическое назначение ревьюверов. (Внутренний API)
    
    ### 3. Code Storage Service (порт 8081)
    Хранение кода в виде деревьев коммитов. 
    Поддержка init, push, checkout, merge. (Внутренний API)
    
    ## Аутентификация
    
    Пользователь идентифицируется через заголовок `X-Username` (имя пользователя).

  version: "1.0.0"
  contact:
    name: API Support
servers:
  - url: http://localhost:8082
    description: User Gateway Service (рекомендуется)
  - url: http://localhost:8080
    description: PR Allocation Service (внутренний)
  - url: http://localhost:8081
    description: Code Storage Service (внутренний)
...
components:
  parameters:
...
    UsernameHeader:
      name: X-Username
      in: header
      required: true
      schema:
        type: string
      description: Имя текущего пользователя (строка)

    # PR Allocation parameters (внутренние)
...
  schemas:
...
    UserProfile:
      type: object
      required: [username, team_name, is_active]
      properties:
        user_id:
          type: string
          description: ID пользователя (в ответе)
        username:
          type: string
        team_id:
          type: string
          format: uuid
          description: UUID команды (в ответе)
        team_name:
          type: string
        is_active:
          type: boolean
...
    PullRequestList:
      type: object
      required: [pull_requests]
      properties:
        pull_requests:
          type: array
          items:
            $ref: '#/components/schemas/GatewayPullRequest'
...
    Team:
      type: object
      required: [team_name, members]
      properties:
        team_id:
          type: string
          format: uuid
          description: UUID команды (в ответе)
        team_name:
          type: string
        members:
          type: array
          items:
            $ref: '#/components/schemas/TeamMember'

    # ========== Internal schemas ==========
...
paths:
...
  /api/me:
    get:
      tags: ["[Gateway] Profile"]
      summary: Получить профиль текущего пользователя
      servers:
        - url: http://localhost:8082
      parameters:
        - $ref: '#/components/parameters/UsernameHeader'
      responses:
        '200':
          description: Профиль пользователя
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/UserProfile'
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ============================================================
  # USER GATEWAY SERVICE - Teams
  # ============================================================
  /api/team/create:
    post:
      tags: ["[Gateway] Teams"]
      summary: Создать команду с участниками
      description: |
        Создаёт команду с указанными участниками.
        Используйте только имена (team_name, username).
      servers:
        - url: http://localhost:8082
      parameters:
        - $ref: '#/components/parameters/UsernameHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [team_name, members]
              properties:
                team_name:
                  type: string
                  description: Уникальное имя команды
                members:
                  type: array
                  items:
                    type: object
                    required: [username]
                    properties:
                      username:
                        type: string
                        description: Уникальное имя пользователя
                      is_active:
                        type: boolean
                        default: true
            example:
              team_name: backend
              members:
                - username: alice
                  is_active: true
                - username: bob
                  is_active: true
      responses:
        '201':
          description: Команда создана
          content:
            application/json:
              schema:
                type: object
                properties:
                  team:
                    $ref: '#/components/schemas/Team'
        '400':
          description: Команда уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/team/get:
    get:
      tags: ["[Gateway] Teams"]
      summary: Получить команду по имени
      servers:
        - url: http://localhost:8082
      parameters:
        - $ref: '#/components/parameters/UsernameHeader'
        - name: team_name
          in: query
          required: true
          schema:
            type: string
          description: Имя команды
      responses:
        '200':
          description: Информация о команде
          content:
            application/json:
              schema:
                type: object
                properties:
                  team:
                    $ref: '#/components/schemas/Team'
        '404':
          description: Команда не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ============================================================
  # USER GATEWAY SERVICE - Repository
  # ============================================================
  /api/repo/init:
    post:
      tags: ["[Gateway] Repository"]
      summary: Инициализировать новый репозиторий
      description: Создаёт root commit с начальным кодом. Используйте имена.
      servers:
        - url: http://localhost:8082
      parameters:
        - $ref: '#/components/parameters/UsernameHeader'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [team_name, repo_name, commit_name, code]
              properties:
                team_name:
                  type: string
                  description: Имя команды
                repo_name:
                  type: string
                  description: Уникальное имя репозитория
                commit_name:
                  type: string
                  description: Имя начального коммита (например "main" или "initial")
                code:
                  type: string
                  format: binary
                  description: ZIP архив с исходным кодом (≤ 100MB)
      responses:
        '201':
          description: Репозиторий инициализирован
          content:
            application/json:
              schema:
                type: object
                properties:
                  commit:
                    $ref: '#/components/schemas/GatewayCommit'
        '403':
          description: Нет доступа
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/repo/push:
    post:
      tags: ["[Gateway] Repository"]
      summary: Создать новый коммит
      description: Создаёт коммит поверх существующего. Используйте имена.
      servers:
        - url: http://localhost:8082
      parameters:
        - $ref: '#/components/parameters/UsernameHeader'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [team_name, repo_name, parent_commit_name, commit_name, code]
              properties:
                team_name:
                  type: string
                  description: Имя команды
                repo_name:
                  type: string
                  description: Имя репозитория
                parent_commit_name:
                  type: string
                  description: Имя родительского коммита
                commit_name:
                  type: string
                  description: Имя нового коммита
                code:
                  type: string
                  format: binary
                  description: ZIP архив с кодом
      responses:
        '201':
          description: Коммит создан
          content:
            application/json:
              schema:
                type: object
                properties:
                  commit:
                    $ref: '#/components/schemas/GatewayCommit'
        '403':
          description: Нет доступа
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/repo/checkout:
    get:
      tags: ["[Gateway] Repository"]
      summary: Получить код коммита
      description: Возвращает ZIP архив с кодом. Используйте имена.
      servers:
        - url: http://localhost:8082
      parameters:
        - $ref: '#/components/parameters/UsernameHeader'
        - name: team_name
          in: query
          required: true
          schema:
            type: string
          description: Имя команды
        - name: repo_name
          in: query
          required: true
          schema:
            type: string
          description: Имя репозитория
        - name: commit_name
          in: query
          required: true
          schema:
            type: string
          description: Имя коммита
      responses:
        '200':
          description: ZIP архив с кодом
          content:
            application/zip:
              schema:
                type: string
                format: binary
        '403':
          description: Нет доступа
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ============================================================
  # USER GATEWAY SERVICE - Pull Requests
  # ============================================================
  /api/pr/create:
    post:
      tags: ["[Gateway] Pull Requests"]
      summary: Создать Pull Request
      description: |
        Создаёт PR и автоматически назначает ревьюверов.
        Все идентификаторы - имена (строки).
      servers:
        - url: http://localhost:8082
      parameters:
        - $ref: '#/components/parameters/UsernameHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, pr_name, team_name, repo_name, source_commit_name, target_commit_name]
              properties:
                title:
                  type: string
                  description: Название PR
                pr_name:
                  type: string
                  description: Уникальное имя PR
                team_name:
                  type: string
                  description: Имя команды
                repo_name:
                  type: string
                  description: Имя репозитория
                source_commit_name:
                  type: string
                  description: Имя source коммита
                target_commit_name:
                  type: string
                  description: Имя target коммита
            example:
              title: "Add user authentication"
              pr_name: "auth-feature"
              team_name: "backend"
              repo_name: "main-service"
              source_commit_name: "feature-branch"
              target_commit_name: "main"
      responses:
        '201':
          description: PR создан
          content:
            application/json:
              schema:
                type: object
                properties:
                  pull_request:
                    $ref: '#/components/schemas/GatewayPullRequest'
        '403':
          description: Нет доступа
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: PR уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/pr/my:
    get:
      tags: ["[Gateway] Pull Requests"]
      summary: Получить мои Pull Requests
      description: PR, где текущий пользователь - автор.
      servers:
        - url: http://localhost:8082
      parameters:
        - $ref: '#/components/parameters/UsernameHeader'
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [OPEN, MERGED, REJECTED]
      responses:
        '200':
          description: Список PR
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PullRequestList'
  /api/pr/reviews:
    get:
      tags: ["[Gateway] Pull Requests"]
      summary: Получить PR на ревью
      description: PR, где текущий пользователь - ревьювер.
      servers:
        - url: http://localhost:8082
      parameters:
        - $ref: '#/components/parameters/UsernameHeader'
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [OPEN, MERGED, REJECTED]
      responses:
        '200':
          description: Список PR на ревью
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PullRequestList'
  /api/pr/approve:
    post:
      tags: ["[Gateway] Pull Requests"]
      summary: Одобрить PR
      description: Одобряет PR и выполняет merge. Только ревьювер.
      servers:
        - url: http://localhost:8082
      parameters:
        - $ref: '#/components/parameters/UsernameHeader'
        - name: team_name
          in: query
          required: true
          schema:
            type: string
          description: Имя команды
        - name: pr_name
          in: query
          required: true
          schema:
            type: string
          description: Имя PR
      responses:
        '200':
          description: PR одобрен и смержен
          content:
            application/json:
              schema:
                type: object
                properties:
                  pull_request:
                    $ref: '#/components/schemas/GatewayPullRequest'
                  merge_commit:
                    $ref: '#/components/schemas/GatewayCommit'
        '403':
          description: Не ревьювер
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: PR не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/pr/reject:
    post:
      tags: ["[Gateway] Pull Requests"]
      summary: Отклонить PR
      description: Отклоняет PR. Только ревьювер.
      servers:
        - url: http://localhost:8082
      parameters:
        - $ref: '#/components/parameters/UsernameHeader'
        - name: team_name
          in: query
          required: true
          schema:
            type: string
          description: Имя команды
        - name: pr_name
          in: query
          required: true
          schema:
            type: string
          description: Имя PR
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        '200':
          description: PR отклонён
          content:
            application/json:
              schema:
                type: object
                properties:
                  pull_request:
                    $ref: '#/components/schemas/GatewayPullRequest'
        '403':
          description: Не ревьювер
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/pr/code:
    get:
      tags: ["[Gateway] Pull Requests"]
      summary: Получить код PR
      description: ZIP с кодом source commit. Для ревью.
      servers:
        - url: http://localhost:8082
      parameters:
        - $ref: '#/components/parameters/UsernameHeader'
        - name: team_name
          in: query
          required: true
          schema:
            type: string
          description: Имя команды
        - name: pr_name
          in: query
          required: true
          schema:
            type: string
          description: Имя PR
      responses:
        '200':
          description: ZIP архив с кодом
          content:
            application/zip:
              schema:
                type: string
                format: binary
        '403':
          description: Нет доступа
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ============================================================
  # PR ALLOCATION SERVICE (Internal)
  # ============================================================
...
  /team/get:
    get:
      tags: ["[PR-Alloc] Teams"]
      summary: "[Internal] Получить команду"
      servers:
        - url: http://localhost:8080
      parameters:
        - $ref: '#/components/parameters/TeamNameQuery'
      responses:
        '200':
          description: Команда
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
...